% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Limit values to a given range}
\usage{
clamp(x, min, max)
}
\arguments{
\item{x}{numeric. Value to validate}

\item{min}{numeric. Minimum value of range.}

\item{max}{numeric. Maximum value of range.}
}
\value{
The clamped value.
}
\description{
This function is useful for input validation. If the value of x is in the given
range then the function returns x. If the value is outside of the range then
the function returns either the max or the min value of the range.
}
\examples{
# Value in range -- returns value
clamp(10, min = 0, max = 100)

# value less than range -- returns min
clamp(-10, min = 0, max = 100)

# value greater than range -- returns max
clamp(110, min = 0, max = 100)
}
