% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{distinct_rgb_palette}
\alias{distinct_rgb_palette}
\title{Generate a distinct RGB color palette}
\usage{
distinct_rgb_palette(n, alpha = 1, ...)
}
\arguments{
\item{n}{numeric. Number of colors to generate}

\item{alpha}{numeric. Transparency level of the color palette (0-1). Default 1.0}

\item{...}{arguments passed to `kmeans()`}
}
\value{
vector of distinct RGB colors
}
\description{
This function uses `kmeans()` over the RGB colorspace to generate N distinct
RGB colors. It's usually not a problem if `kmeans()` does not converge (colors
are distinct enough for most purposes). If you get warnings, you can modify
the default settings of the `kmeans()` function by passing in additional
arguments. For example, increasing iterations can be done by passing
`iter.max = 100`. Changing the kmeans algorithm can be done by specifying
`algorithm = "Lloyd"`. Changing these arguments may eliminate warnings.
}
