% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.col2assay}
\alias{.col2assay}
\title{Helper function for converting columns to assay data}
\usage{
.col2assay(df, rows, cols, vals)
}
\description{
Helper function for converting columns to assay data
}
