% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{dsc}
\alias{dsc}
\title{Calculate the Dispersion Separability Criterion (DSC)}
\usage{
dsc(x, g)
}
\arguments{
\item{x}{A numeric matrix (features in rows, samples in columns)}

\item{g}{A vector containing the batch assignment for each sample}
}
\value{
DSC metric
}
\description{
This function computes the \href{https://ibl.mdanderson.org/public-software/tcga-batch-effects/}{Dispersion Separability Criterion (DSC)}
metric which was defined by the MD Anderson Bioinformatics team for computing the strength of
batch effects in high-throughput sequencing data from TCGA.
}
\examples{
set.seed(123)

# Small DSC when samples drawn from same distribution
x1 <- matrix(rnorm(500 * 100), nrow = 500, ncol = 100)
batches <- rep(c("Batch_1", "Batch_2"), each = 50)
dsc(x = x1, g = batches)

# Large DSC when samples have batch effect present
x2 <- x1
x2[, 51:100] <- x2[, 51:100] + 50
dsc(x2, batches)

}
