% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env2global}
\alias{env2global}
\title{Extract variable from an environment and remove that environment}
\usage{
env2global(x, remove = TRUE)
}
\arguments{
\item{x}{Environment to extract variables from}

\item{remove}{Should the Environment be removed after extracting variables? Default TRUE}
}
\value{
variables from x are assigned to the GlobalEnv after execution
}
\description{
This function will extract all variables from the given environment and
assign them to the global environment and then optionally remove the
environment.
}
\examples{
my_env <- new.env()
my_env$X <- 1000
my_env$Y <- 1:10

# Extract X and Y to global environment and remove my_env
env2global(my_env)

# Extract X and Y to global environment and keep my_env
env2global(my_env, remove = FALSE)

}
