% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{exact_cor_test}
\alias{exact_cor_test}
\title{Exact Permutation Test}
\usage{
exact_cor_test(x, y, cor_method)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of "pearson", "kendall", or "spearman": can be abbreviated.}
}
\value{
numeric vector. Correlations from permutations of x with y
}
\description{
In the case where the number of number of permutations chosen is greater
than the actual number of permutations, perform an exact test instead.
The code for getting all distinct permutations was stolen from:
https://stackoverflow.com/a/34287541
}
\keyword{internal}
