% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean of a vector}
\usage{
geometric_mean(x, zero_propagate = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector of non-negative values}

\item{zero_propagate}{logical. Should zeros be included in the calculation. Default FALSE.}

\item{na.rm}{logical. Remove NAs from calculation}
}
\value{
geometric mean of the vector
}
\description{
Calculate the geometric mean of a vector
Stolen from \url{https://stackoverflow.com/a/25555105}.
}
\examples{
x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 1500)
geometric_mean(x)
}
