% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-axis-limits.R
\name{get_axis_limits}
\alias{get_axis_limits}
\title{Get the axis limits of a ggplot2 object}
\usage{
get_axis_limits(p)
}
\arguments{
\item{p}{ggplot plot object}
}
\value{
list with elements for x and y limits
}
\description{
Internal helper function used to get the axis limits of a plot
}
\keyword{internal}
