% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE161650_de}
\alias{GSE161650_de}
\title{Differential expression results from GSE161650}
\format{
A data frame with 11,002 rows and 6 columns:
\describe{
  \item{feature_id}{Gene Symbol}
  \item{logFC}{log2 fold-change THZ1 vs DMSO}
  \item{unshrunk.logFC}{unshrunk log2 fold-change THZ1 vs DMSO}
  \item{logCPM}{Average logCPM across all samples}
  \item{PValue}{Uncorrected PValue}
  \item{FDR}{FDR Adjusted PValue}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE161650>
}
\usage{
GSE161650_de
}
\description{
A subset of differential expression results comp THZ1 (n=3) vs DMSO (n=3) 
samples after analysis with edgeR. See Details for citation.
}
\details{
Jiang B, Gao Y, Che J, Lu W et al. Discovery and resistance mechanism of a selective CDK12 degrader. 
Nat Chem Biol 2021 Jun;17(6):675-683. PMID: 33753926
}
\keyword{datasets}
