% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE161650_lc}
\alias{GSE161650_lc}
\title{Normalized log2 counts from GSE161650}
\format{
A matrix with 11,002 rows (Gene Symbols) and 6 columns (Samples).
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE161650>
}
\usage{
GSE161650_lc
}
\description{
A subset of log2 expression counts containing THZ1 (n=3) vs DMSO (n=3)
samples after TMM normalization and log2 transformation. See Details for
citation
}
\details{
Jiang B, Gao Y, Che J, Lu W et al. Discovery and resistance mechanism of a selective CDK12 degrader.
Nat Chem Biol 2021 Jun;17(6):675-683. PMID: 33753926
}
\keyword{datasets}
