% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath-age.R
\name{horvath_age}
\alias{horvath_age}
\title{Transform ages using Horvath's method}
\usage{
horvath_age(x, adult_age = 20)
}
\arguments{
\item{x}{numeric vector of ages}

\item{adult_age}{Age of the adult of the species. From the original manuscript, 
20 for humans and 15 for chimpanzees.}
}
\value{
transformed ages
}
\description{
Transform age in years using the log transformation defined in Horvath (2013)
}
\details{
\href{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115}{Horvath, S. DNA methylation age of human tissues and cell types. 
Genome Biol 14, 3156 (2013). https://doi.org/10.1186/gb-2013-14-10-r115}
}
\examples{

plot(\(x) coriell::horvath_age(x), from=0, to=100, xlab="Age", ylab="Transformed Age")
}
