% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute}
\alias{impute}
\title{Perform simple imputation on rows of a matrix}
\usage{
impute(x, fun = median)
}
\arguments{
\item{x}{numeric matrix or data.frame that can be converted to a numeric matrix}

\item{fun}{Imputation function to apply to rows of the matrix. Default median}
}
\value{
Matrix with imputed values
}
\description{
Impute NA values for each row of a matrix.
}
\examples{
# Create a matrix of values with NAs
X <- matrix(runif(25), 5, dimnames = list(paste0("CpG", 1:5), paste0("Sample", 1:5)))
X[sample.int(25, 5)] <- NA
X

# Impute missing values with row medians
impute(X)

# Impute missing values with arbitrary function
impute(X, fun = function(x) { -1} )
}
