% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lerp}
\alias{lerp}
\title{Linear interpolation of a value}
\usage{
lerp(x, min, max)
}
\arguments{
\item{min}{numeric. Minimum of the desired range.}

\item{max}{numeric. Maximum of the desired range.}

\item{x.}{numeric. Normalized input value to be transformed. Value must be between 0-1.}
}
\value{
Value of x within the desired range given by min, max.
}
\description{
Linear interpolation of a value
}
\examples{
lerp(0.25, min = 0, max = 100)
}
