% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-to-matrix.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{List to Matrix}
\usage{
list_to_matrix(sets)
}
\arguments{
\item{sets}{named list of sets (vectors) to be converted to binary matrix.}
}
\value{
binary matrix with a column for each set. rownames of the matrix represent the union of all sets.
A '1' indicates the inclusion of the element in the set.
}
\description{
Convert a list of sets into a binary matrix.
}
\examples{
sets <- list(
  "set1" = letters[1:5],
  "set2" = letters[2:6],
  "set3" = letters[1:7]
)

list_to_matrix(sets)
}
