% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_value}
\alias{map_value}
\title{Map a value in one range to a value in another}
\usage{
map_value(x, old_min, old_max, new_min, new_max)
}
\arguments{
\item{x}{numeric. Value to be transformed}

\item{old_min}{numeric. Minimum value of source range.}

\item{old_max}{numeric. Maximum value of source range.}

\item{new_min}{numeric. Minimum value of desired range.}

\item{new_max}{numeric. Maximum value of desired range.}
}
\value{
Value of x mapped to the desired range.
}
\description{
Map a value in one range to a value in another
}
\examples{
map_value(50, old_min = 0, old_max = 100, new_min = 100, new_max = 1000)
}
