% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_de}
\alias{meta_de}
\title{Combine p-values across multiple differential expression studies}
\usage{
meta_de(
  l,
  p_method = "fisher",
  lfc_threshold = 0,
  pval_col = "p.value",
  lfc_col = "logFC",
  feature_col = "gene_id",
  ...
)
}
\arguments{
\item{l}{List of data.frames containing differential expression results. 
Every data.frame in the list must have the same columns names.}

\item{p_method}{p-value combination method. One of c("simes", "holm-min", 
"berger", "fisher", "pearson", "wilkinson", "stouffer"). Default "fisher".}

\item{lfc_threshold}{numeric scalar defining the threshold at which an effect 
is "up" or "down"}

\item{pval_col}{character string of the column name in each data.frame that 
contains the p-values to combine. Default "p.value"}

\item{lfc_col}{character string of the column name in each data.frame that 
contains the log-fold change values to combine. Default "logFC"}

\item{feature_col}{character string of the column name in each data.frame 
that contains the feature names (i.e. genes) over which to combine. Default 
"gene_id"}

\item{...}{Additional arguments passed to \code{metapod::combineParallelPValues()}}
}
\value{
data.frame with columns containing the combined p-values, the 
summarized direction of the effect, and additional summary statistics about 
combined logFC values for each gene.
}
\description{
This function provides an interface to p-value combination methods 
implemented by the \code{metapod} package for use in the context of 
differential expression.
}
\details{
The "n.logFC" column of the resulting data.frame reports the number of non-NA
logFC values. The "n.pos.logFC" and the "n.neg.logFC" columns report the 
number of logFC values across experiments that exceed the value given by the
"lfc_threshold" argument. All other summary statistics on logFC values are 
performed after removing NA values.
}
\examples{
df1 <- data.frame(
  gene_id = c("A", "B", "C", "D", "E"),
  logFC = c(-1, 0, 1.5, 3.2, 0.25),
  p.value = c(0.015, 0.1, 0.000005, 0.006, 0.9)
)

df2 <- data.frame(
  gene_id = c("A", "B", "C", "D", "F"),
  logFC = c(-1.5, 0.1, 1.6, 2.9, -2.1),
  p.value = c(0.0006, 0.3, 0.000058, 0.0005, 0.0087)
)

df3 <- data.frame(
  gene_id = c("G", "B", "C", "D", "F"),
  logFC = c(2.6, 0.2, 1.25, 3.3, -2.7),
  p.value = c(0.085, 0.5, 0.005, 0.0006, 0.00007)
)

# Note: genes do not have to be present across all studies 
L <- list(exp1=df1, exp2=df2, exp3=df3)

meta_de(L)
}
