% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_pcombine}
\alias{meta_pcombine}
\title{Perform p-value combination on differential expression results}
\usage{
meta_pcombine(
  exp_list,
  lfc_col = "logFC",
  pval_col = "FDR",
  gene_col = "feature_id",
  all_common = FALSE,
  lfc_fun = mean,
  method = c("fisher", "pearson", "stouffer", "tippet", "wilkinson"),
  plot = FALSE,
  plot_lfc = 0,
  plot_pval = 0.05,
  plot_labels = FALSE,
  plot_count = TRUE,
  ...
)
}
\arguments{
\item{exp_list}{list() of data.frames of differential expression results.}

\item{lfc_col}{Column name in the expression data.frames that contains log fold-change values. default = "logFC".}

\item{pval_col}{Column name in expression data.frames that contains the significance values. default = "FDR".}

\item{gene_col}{Column name in the expression data.frames that contains the gene IDs. default = "feature_id".}

\item{all_common}{Logical. Use only genes present in all experiments. default = FALSE.}

\item{lfc_fun}{Function for summarizing logFC values across studies.}

\item{method}{Method for combining p-values. One of c("fisher", "pearson", "stouffer", "tippet", "wilkinson"). default "fisher".}

\item{plot}{Logical. Display a volcano plot using the combined values. default = FALSE}

\item{plot_lfc}{Numeric. LogFC cutoff used when plotting. default = 0}

\item{plot_pval}{Numeric. P-value cutoff used when plotting. default = 0.05.}

\item{plot_labels}{Logical. Display labels for significant genes on plot. default = FALSE}

\item{plot_count}{Logical. Display count of up/down genes on plot. default = TRUE}

\item{...}{Additional arguments passed to lfc_fun.}
}
\value{
data.table with columns containing the combined logFC and p-values.
}
\description{
This function aims to combine p-values and log fold-change values into a single
combined p-value and logFC value for each gene across all studies. Several methods
for combining p-values are included. The useer can also pass an arbitrary function
to the \code{lfc_fun} argument used to summarize the logFC values across studies.
}
\details{
The following p-value combination methods are available:
\itemize{
 \item fisher: Uses the sum of the logarithms for the p-values. Sensitive to very small p-values therefore a single significant study can lead to a very small combined p-value.
 \item pearson: Similar to Fisher's method but sensitive to large p-values; therefore more false negatives are obtained.
 \item stouffer: (not yet implemented) Recommended when weights for each study can be calculated.
 \item tippet: Uses the minimum of the p-values across all studies. Recommended when the aim is to discard genes.
 \item wilkinson: Uses the maximum of the p-values across all studies. Recommended when the aim is to identify the most robust genes.
}
}
