% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_vote}
\alias{meta_vote}
\title{Perform vote counting on differential expression results}
\usage{
meta_vote(
  exp_list,
  lfc_col = "logFC",
  pval_col = "FDR",
  gene_col = "feature_id",
  lfc = 0,
  pval = 0.05,
  meta_prop = 0.5,
  all_common = FALSE,
  plot = FALSE
)
}
\arguments{
\item{exp_list}{list() of data.frames of differential expression results.}

\item{lfc_col}{Column name in expression data.frames that contains the log fold-change estimates. default = "logFC".}

\item{pval_col}{Column name in expression data.frames that contains the significance values. default = "FDR".}

\item{gene_col}{Column name in the expression data.frames that contains the gene IDs. default = "feature_id".}

\item{lfc}{Numeric. Log fold-change value used as cutoff for determining differential expression in each study. default = 0.}

\item{pval}{Numeric. Significance value used as cutoff for determining differential expression in each study}

\item{meta_prop}{Numeric. Proportion of studies a gene must be differentially expressed in to be considered as 'meta-differentially expressed'. default = 0.5.}

\item{all_common}{Logical. Use only genes present in all experiments. default = FALSE.}

\item{plot}{Logical. Display a metavolcano ggplot2 plot of the results. default = FALSE.}
}
\value{
data.table of meta-expression results with the following columns:
\itemize{
  \item gene_col: Genes included in the final voting procedure. 
  \item n_de: The total number of studies in which the gene was differentially expressed (either up or down-regulated). 
  \item prop_de: The proportion of studies in which the gene was differentially expressed (either up or down-regulated).
  \item sign_consistency: The number of studies in which a gene was up-regulated minus the number of studies in which a gene was down-regulated.
  \item vote: The final vote for the gene. based on \code{sign(sign consistency) & abs(sign consistency) >= meta_prop * N studies}.
}
}
\description{
Perform vote-counting on a list of differential expression results. This method 
uses user defined p-value and log fold-change cutoffs to determine the number
of times a gene is differentially expressed across multiple studies. A gene is
then considered 'meta-differentially expressed' if it is differentially expressed
in the user defined proportion of studies.
}
\details{
There are certain caveats to consider when using the vote counting method. 1.
combining FDR and logFC cutoffs after the fact destroys FDR correction. See
\href{https://academic.oup.com/bioinformatics/article/25/6/765/251641}{McCarthy & Smyth, 2009} for more details. 
2. The direction of expression is not considered when the final vote is cast. For example, if 
geneA is up-regulated in 4 studies but down-regulated in 1 then the final vote for 
geneA will be 'up-regulated'.
}
