% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minmax}
\alias{minmax}
\title{Min-Max normalize a value within a given range}
\usage{
minmax(x, min = NA, max = NA)
}
\arguments{
\item{x}{numeric. Value to be transformed (normalized).}

\item{min}{numeric. Minimum value of range. Default NA, use the min of the supplied vector}

\item{max}{numeric. Maximum value of range. Default NA, use the max of the supplied vector}
}
\value{
Normalized value (0-1) of x in the range given by min, max.
}
\description{
Min-Max normalize a value within a given range
}
\examples{
minmax(c(0, 1, 10, 100))

# Use a scale outside of the range of x
minmax(c(0, 1, 10, 100), min = 0, max = 1000)
}
