% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize a value within a given range}
\usage{
normalize(x, min, max)
}
\arguments{
\item{x}{numeric. Value to be transformed (normalized).}

\item{min}{numeric. Minimum value of range.}

\item{max}{numeric. Maximum value of range.}
}
\value{
Normalized value (0-1) of x in the range given by min, max.
}
\description{
Normalize a value within a given range
}
\examples{
normalize(10, min = 0, max = 100)
}
