% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_by_iqr}
\alias{outliers_by_iqr}
\title{Flag outliers by IQR}
\usage{
outliers_by_iqr(x, threshold = 1.5, direction = c("both", "low", "high"))
}
\arguments{
\item{x}{numeric vector}

\item{threshold}{threshold for the upper and lower limits, A.K.A. Tukey's fences}

\item{direction}{return TRUE if the value is above or below the outlier cutoff.
default "both", samples above and below the threshold are called outliers.}
}
\value{
boolean vector indicating which values of the input vector are flagged as outliers
}
\description{
Flag outliers based on the interquartile range
}
\examples{

x <- c(1, 1, 2, 2, 4, 6, 11)

outliers_by_iqr(x)

# Using direction="low" disregards outliers above threshold, for example
outliers_by_iqr(x, direction="low")

}
