% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_by_mad}
\alias{outliers_by_mad}
\title{Flag outliers by MAD}
\usage{
outliers_by_mad(x, threshold = 3, direction = c("both", "low", "high"))
}
\arguments{
\item{x}{numeric vector}

\item{threshold}{threshold for number of MADs used to determine outlier. default 3}

\item{direction}{return TRUE if the value is above or below the outlier cutoff. 
default "both", samples above and below the threshold are called outliers.}
}
\value{
boolean vector indicating which values of the input vector are flagged as outliers
}
\description{
Flag outliers based on the median absolute deviation
}
\examples{

x <- c(1, 1, 2, 2, 4, 6, 11)

outliers_by_mad(x)

# Using direction="low" disregards outliers above threshold, for example
outliers_by_mad(x, direction="low")

}
