% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_by_z}
\alias{outliers_by_z}
\title{Flag outliers by z-score}
\usage{
outliers_by_z(x, threshold = 3, direction = c("both", "low", "high"))
}
\arguments{
\item{x}{numeric vector}

\item{threshold}{threshold for number of SDs used to determine outlier. default 3}

\item{direction}{return TRUE if the value is above or below the outlier cutoff. 
default "both", samples above and below the threshold are called outliers.}
}
\value{
boolean vector indicating which values of the input vector are flagged as outliers
}
\description{
Flag outliers that exceed a given z-score value.
}
\examples{
x <- c(rnorm(10), 100)

outliers_by_z(x)

# Using direction="low" disregards outliers above threshold, for example
outliers_by_z(x, direction="low")

}
