% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-fisher-test.R
\name{pairwise_fisher_test}
\alias{pairwise_fisher_test}
\title{Perform pairwise fisher.tests on an input matrix}
\usage{
pairwise_fisher_test(x, ref)
}
\arguments{
\item{x}{numeric 2 x m data.frame, matrix, or table}

\item{ref}{column name of the reference column of the input matrix}
}
\value{
data.frame of results for each test.
}
\description{
This function will perform a \code{fisher.test()} for all columns relative
to the given reference column of the input 2 x m matrix and return a
data.frame of the results.
}
\details{
The function performs a fisher test for each test column relative to
the reference column of the input matrix. The input matrix must be a 2 x m
matrix. It is the user's responsibility to ensure the levels of the rows align
with the desired odds ratio calculation. e.g. R will calculate the odds ratio
as \code{(x[1, "test_col"] / x[2, "test_col"]) / (x[1, "ref_col"] / x[2, "ref_col"])}.
It is the user's responsibility to make sure \code{x[1, ]} and \code{x[2, ]}
are in the desired order.
}
\examples{
# From chisq.test docs:
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(
  gender = c("F", "M"),
  party = c("Democrat", "Independent", "Republican")
)
pairwise_fisher_test(M, ref = "Independent")
}
