% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_intersections}
\alias{pairwise_intersections}
\title{Get unique pairwise intersections of a list of vectors}
\usage{
pairwise_intersections(x)
}
\arguments{
\item{x}{List of vectors to perform intersections on}
}
\value{
data.table

a data.table containing columns for the sets being compared, a list column
which contains the actual values in the intersection, a column with the 
intersection size, and a column with the Jaccard index.
}
\description{
This function takes in a list of vectors and performs pairwise set 
intersections for all unique pairs of vectors in the list.
}
\examples{
l <- list(
  Set1 = c("A", "B", "C"), 
  Set2 = c("B", "C", "D"),
  Set3 = c("X", "Y", "Z"),
  Set4 = LETTERS
)

pairwise_intersections(l)
}
