% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_testing.R
\name{perm_cor}
\alias{perm_cor}
\title{Correlation Permutation Test Function}
\usage{
perm_cor(x, y, n_perm, method = "spearman")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{n_perm}{numeric. The number of permutations to perform}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of "pearson", "kendall", or "spearman" (default): can be abbreviated.}
}
\value{
numeric vector of length n_perm
}
\description{
Internal function for performing a correlation permutation test
}
