% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_correlation_test}
\alias{permutation_correlation_test}
\title{Perform a permutation correlation test on every row of a matrix}
\usage{
permutation_correlation_test(X, y, n_perm = 10000, n_core = 1, ...)
}
\arguments{
\item{X}{numeric matrix or data.frame that can be converted to a numeric matrix}

\item{y}{numeric vector of values to correlate with rows of X}

\item{n_perm}{integer. The desired number of permutations to sample from. Default (10,000)}

\item{n_core}{integer. The number of cores to use for processing. Default (1)}

\item{...}{Additional arguments to pass to `cor` function}
}
\description{
Compute a correlation value for every row of X against the vector y and n 
random permutations of y. If the number of possible permutations is less than
the the argument n_perm then an exact test is performed instead. In both 
cases the function returns a data.frame of the original data with additional
columns for the test statistic, empirical p-value, and FDR corrected empirical
p-value.
}
\examples{
{
# generate example data
X <- matrix(runif(1e3 * 10), nrow = 1e3, ncol = 10)
y <- 1:10

# correlate rows of X with 1,000 random permutations of vector y
res <- coriell::permutation_correlation_test(X, y, n_perm = 1e3, n_core = 8, method = "spearman")
}
}
