% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutations}
\alias{permutations}
\title{Generate all permutations of a vector}
\usage{
permutations(x)
}
\arguments{
\item{x}{vector}
}
\value{
factorial(x) x length(x) matrix
}
\description{
Given the input vector, generate a matrix of permutations where each row
represents a permutation of the data. Stolen from:
https://stackoverflow.com/a/34287541
}
\examples{
v <- letters[1:4]
p <- permutations(v)
p
}
