% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\alias{plot_boxplot.default}
\alias{plot_boxplot.matrix}
\alias{plot_boxplot.data.frame}
\alias{plot_boxplot.SummarizedExperiment}
\title{Show boxplots for columns of data in a matrix}
\usage{
plot_boxplot(x, ...)

\method{plot_boxplot}{default}(x)

\method{plot_boxplot}{matrix}(
  x,
  metadata = NULL,
  fillBy = NULL,
  rle = FALSE,
  violin = FALSE,
  ...
)

\method{plot_boxplot}{data.frame}(
  x,
  metadata = NULL,
  fillBy = NULL,
  rle = FALSE,
  violin = FALSE,
  ...
)

\method{plot_boxplot}{SummarizedExperiment}(
  x,
  assay = "counts",
  fillBy = NULL,
  rle = FALSE,
  violin = FALSE,
  ...
)
}
\arguments{
\item{x}{matrix of expression values or \code{SummarizedExperiment} object}

\item{...}{Additional parameters passed to \code{ggplot2::geom_boxplot()} or
\code{ggplot2::geom_violin()}}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata
must match the colnames of the input matrix. Default NULL, each sample in x
will be plotted individually.}

\item{fillBy}{metadata column used to fill boxplots. Default NULL,
each sample will be a distinct color.}

\item{rle}{Plot the relative log expression value. This option assumes
that the input matrix has already been logged. Be sure to take the log of the
input matrix prior to setting this option.}

\item{violin}{Plot the data as a violin plot instead of a boxplot. Default FALSE.}

\item{assay}{If a SummarizedExperiment is supplied what assay is used. Default "counts"}
}
\value{
ggplot object
}
\description{
Create a boxplot (or violin plot) of distributions for the given matrix. Optionally plot the
relative log expression of the matrix.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the boxplot by sample
plot_boxplot(GSE161650_lc)

# Plot the boxplot by coloring each Group
plot_boxplot(GSE161650_lc, metadata, fillBy = "Group")

# Create a violin plot after RLE transformation
plot_boxplot(GSE161650_lc, metadata, fillBy = "Group", rle = TRUE, violin = TRUE)
}
