% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density.R
\name{plot_density}
\alias{plot_density}
\alias{plot_density.default}
\alias{plot_density.matrix}
\alias{plot_density.data.frame}
\alias{plot_density.SummarizedExperiment}
\title{Show density distributions for columns of data in a matrix}
\usage{
plot_density(x, ...)

\method{plot_density}{default}(x)

\method{plot_density}{matrix}(x, metadata = NULL, colBy = NULL, ...)

\method{plot_density}{data.frame}(x, metadata = NULL, colBy = NULL, ...)

\method{plot_density}{SummarizedExperiment}(x, assay = "counts", colBy = NULL, ...)
}
\arguments{
\item{x}{gene by sample matrix or \code{SummarizedExperiment} object}

\item{...}{Additional parameters passed to \code{ggplot2::geom_density()}}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata
must match the colnames of the input matrix. Default NULL, each sample in the
matrix will be plotted.}

\item{colBy}{metadata column used to color density lines. Default NULL,
each sample in the matrix will be plotted.}

\item{assay}{assay of the SummarizedExperiment object. Default "counts"}
}
\value{
ggplot object
}
\description{
Create a density plot of values for columns in the given expression matrix.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the density by sample
plot_density(GSE161650_lc)

# Color each sample by their Group in metadata
plot_density(GSE161650_lc, metadata, colBy = "Group")

}
