% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{plot_metavolcano}
\alias{plot_metavolcano}
\title{Create Meta-volcano plot}
\usage{
plot_metavolcano(meta_df, label_col)
}
\arguments{
\item{meta_df}{data.table of meta_vote results.}
}
\value{
ggplot2 object
}
\description{
Specialty function for creating metavolcano plots.
}
