% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-meth-hist.R
\name{plot_meth_hist}
\alias{plot_meth_hist}
\title{Create a binned histogram of methylation percentage statistics}
\usage{
plot_meth_hist(
  x,
  bar_color = "cornflowerblue",
  border_color = "white",
  plot_title = NULL,
  y_label_cex = 1,
  x_label_cex = 1,
  label_cex = 0.8,
  y_axis_cex = 0.8,
  x_axis_cex = 0.8,
  y_axis_label = "Number CpGs"
)
}
\arguments{
\item{x}{numeric vector of methylation beta-values}

\item{bar_color}{color of the histogram bars. Default = 'cornflowerblue'}

\item{border_color}{color of the borders for each bar. Default = 'white'}

\item{plot_title}{title of the plot. Default = NULL}

\item{y_label_cex}{size of the y-axis label. Default = 1}

\item{x_label_cex}{size of the x-axis label. Default = 1}

\item{label_cex}{size of the barplot labels. Default = 0.8}

\item{y_axis_cex}{size of the y-axis values. Default = 1}

\item{x_axis_cex}{size of the x-axis values. Default = 0.8}

\item{y_axis_label}{name of the y-axis. Default = 'Number of CpGs'}
}
\value{
histogram with methylation statistics
}
\description{
Create a binned histogram of methylation percentage statistics
}
\examples{

# Example methylation data (single column of a matrix)
data <- rbeta(runif(1e4), 0.6, 0.8)
plot_meth_hist(data, plot_title="Percentage Methylated CpGs per Binned beta-value")

# Plot all columns of a matrix
m <- replicate(6, rbeta(1e4, 0.5, 0.8))
colnames(m) <- paste("Sample", 1:6)

# Create a square grid
n <- ncol(m)
ncols <- ceiling(sqrt(n))
nrows <- ceiling(n / ncols)

par(mfrow = c(nrows, ncols))
for (i in 1:n) {
  plot_meth_hist(as.numeric(m[, i]), plot_title = colnames(m)[i], x_label_cex = 0.8)
}

}
