% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-quant-file.R
\name{process_quant_file}
\alias{process_quant_file}
\title{Process a quant file from REdiscoverTE pipeline}
\usage{
process_quant_file(quant_file, gene_annot, rmsk_annot)
}
\arguments{
\item{quant_file}{path to the quant file to be processed}

\item{gene_annot}{REdiscoverTE gene annotation data.frame}

\item{rmsk_annot}{REdiscoverTE RepeatMasker annotation data.frame}
}
\value{
list containing two data.tables. One for the gene counts ("gene") and the other for repeat counts ("re")
}
\description{
Specialty function for processing quants files produced by the REdiscoverTE
pipeline. The function takes a quant file as input and calculates the counts
per gene and per repetitive element subfamily and returns both count data.frames
in a list. The user must provide the pre-processed gene and repeatmasker
annotations.
}
\examples{
\dontrun{
# read in annotation files
gencode <- as.data.table(readRDS("/mnt/data/gdata/human/REdiscoverTE_hg38/REdiscoverTE_hg38_GFP/GENCODE.V26.Basic_Gene_Annotation_md5_GFP.rds"))
rmsk <- as.data.table(readRDS("/mnt/data/gdata/human/REdiscoverTE_hg38/rmsk_annotation.RDS"))

# specify path to quant file
my_quant_file <- "/path/to/quant.sf"

result <- process_quant_file(my_quant_file, gencode, rmsk)

# If you have a list of quant files, you can apply over all with multiple cores
my_quants <- list.files("path/to/quants_dir", pattern = "quant.sf", full.names = TRUE)

processed <- parallel::mclapply(
  X = my_quants,
  FUN = process_quant_file,
  gene_annot = gencode,
  rmsk_annot = rmsk,
  mc.cores = 12
)
}
}
