% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank-threshold.R
\name{rank_threshold}
\alias{rank_threshold}
\title{Calculate threshold value on ranked input}
\usage{
rank_threshold(x, show = FALSE)
}
\arguments{
\item{x}{numeric vector of values.}

\item{show_plot}{logical. default (FALSE). Show the scatter plot for the given distribution.}
}
\value{
named vector containing "x" and "y" position of calculated threshold
}
\description{
Calculate a threshold value of a vector based on the unimodal threshold method
described by \href{https://users.cs.cf.ac.uk/Paul.Rosin/resources/papers/unimodal2.pdf"}{Rosin 2001}
for image thresholding. Conceptually, the method attempts to draw a line from
the peak of the distribution to the tail and calculates the maximum distance
from a point to that line. The point where the distance is maximized is the
threshold.
}
\details{
The function takes a numeric vector and ranks the values using \code{data.table::frank}.
with \code{ties.method = "first"}. The returned values for the threshold "x" and "y",
indicate the rank of the data point and the value at that rank, respectively.
}
\examples{
# simulate exponential counts
x <- (1:100)^exp(3)
rank_threshold(x, show = TRUE)
}
