% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-helpers.R
\name{remove_components}
\alias{remove_components}
\title{Remove principle components from data}
\usage{
remove_components(x, components = 1, ...)
}
\arguments{
\item{x}{data.frame or matrix of original data.}

\item{components}{numeric vector of components to remove from original data. Default (1)}

\item{...}{Additional arguments passed to \code{prcomp()}}
}
\value{
data.frame of values in the original units after removing components
}
\description{
Reconstruct data after removing principle components. This function will 
reconstruct a truncated version of the original data matrix after removing
the specified principle components. The source code for this function was 
stolen from \href{https://stats.stackexchange.com/questions/57467/how-to-perform-dimensionality-reduction-with-pca-in-r/57478#57478}{stack exchange}
}
\examples{
# Remove first two components from dataset
trunc <- remove_components(USArrests, components = 1:2, scale. = TRUE, center = TRUE)

# original data
head(USArrests)

# reconstructed data
head(trunc)
}
