% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_var}
\alias{remove_var}
\alias{remove_var.default}
\alias{remove_var.matrix}
\alias{remove_var.data.frame}
\alias{remove_var.SummarizedExperiment}
\alias{remove_var.DelayedArray}
\title{Remove low variance features from a matrix}
\usage{
remove_var(x, ...)

\method{remove_var}{default}(x)

\method{remove_var}{matrix}(x, p)

\method{remove_var}{data.frame}(x, p)

\method{remove_var}{SummarizedExperiment}(x, p, assay = "counts")

\method{remove_var}{DelayedArray}(x, p)
}
\arguments{
\item{x}{matrix, numeric data.frame, or SummarizedExperiment object}

\item{p}{proportion of low variance features to remove}

\item{assay}{if SummarizedExperiment, what assay to use. Default = "counts"}
}
\value{
matrix with lowest variance features removed
}
\description{
This function removes the p lowest variance features from a matrix. The 
function expects features in rows and samples in columns (e.g. an expression
matrix).
}
\examples{

# Remove 80\% lowest variance features
removed <- remove_var(GSE161650_lc, p = 0.8)
}
