% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{sample_n_random_cor}
\alias{sample_n_random_cor}
\title{Sample Random Correlations from a dataframe}
\usage{
sample_n_random_cor(df, y, n = 1e+06, cor_method = "spearman")
}
\arguments{
\item{df}{data.frame. Site by Samples data.frame}

\item{y}{numeric vector. Numeric vector of values used to correlate with each row of df}

\item{n}{integer. Number of random correlations to return. Default (10000)}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of "pearson", "kendall", or "spearman" (default): can be abbreviated.}
}
\value{
named numeric vector of correlation values and row indeces. Vector can contain NAs if row variance == 0.
}
\description{
Selects n random rows from a dataframe with replacement. For each random row,
permute vector y and perform correlation with the given cor_method.
}
\examples{
\dontrun{
library(methylKit)

ages <- data.frame(age = c(30, 80, 34, 30, 80, 40))

sim_meth <- dataSim(
  replicates = 6,
  sites = 1000,
  treatment = c(rep(1, 3), rep(0, 3)),
  covariates = ages,
  sample.ids = c(paste0("test", 1:3), paste0("ctrl", 1:3))
)

# extract the methylation as percentages and coerce to data.frame
perc_meth <- as.data.frame(percMethylation(sim_meth))

# Get 1_000_000 random correlations from perc_meth data.frame
cors <- sample_n_random_cor(perc_meth, y = ages$age)

# histogram of distribution -- filter NAs if present
hist(cors[!is.na(cors)])
}
}
