% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-counts.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Generate simulated RNA-seq data for testing purposes}
\usage{
simulate_counts(
  n_genes = 1000,
  n_up = 50,
  n_down = 50,
  n_samples = 6,
  groups = c("ctl", "trt"),
  de_group = "trt",
  mu = 10,
  phi = 0.1,
  count_offset = 10
)
}
\arguments{
\item{n_genes}{integer. The number of gene to simulate}

\item{n_up}{integer. The number of genes the count offset will be added to}

\item{n_down}{integer. the number of genes the count offset will be subtracted from. Negative numbers will be coerced to zeros.}

\item{n_samples}{integer. The number of samples to simulate}

\item{groups}{character vector. Character vector specifying the names of the groups to simulate}

\item{de_group}{character. The name of the group to be differentially expressed}

\item{mu}{numeric. Mean value of counts}

\item{phi}{numeric. Dispersion parameter}

\item{count_offset}{numeric. Count offset to be applied to the up/down genes.}
}
\value{
list containing the count matrix and a vectors indicating which rows were modified.
}
\description{
Simulate an RNASeq count matrix. Adapted from edgeR documentation.
This function is not meant for any serious modelling. It's really used mostly for testing
plotting and summarizing functions internally. It's good for generating test data.
}
\examples{
# simulate counts using default parameters
sim <- simulate_counts()

# extract the count matrix from the simulation result
counts <- sim$table

# show rows where counts were modified
counts[sim$de_genes, ]
}
