% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample-counts.R
\name{subsample_counts}
\alias{subsample_counts}
\title{Sub-sample a count matrix to the smallest library size}
\usage{
subsample_counts(counts, seed = 123)
}
\arguments{
\item{counts}{numeric matrix of integer counts.}

\item{seed}{integer. Seed for random number generator. Default (123)}
}
\value{
numeric matrix of sub-sampled counts.
}
\description{
Randomly sub-sample a count matrix so that the column sums are equivalent
to the library size of the smallest sample (column). Returns a numeric matrix
of subsampled counts with the same dimensions as the input matrix.
}
\examples{
library(coriell)

# generate simulated count data
mat <- simulate_counts()$table
colSums(mat)

ss <- subsample_counts(mat)
colSums(ss)
}
