% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-dge.R
\name{summarize_dge}
\alias{summarize_dge}
\title{Summarize RNA-seq expression results}
\usage{
summarize_dge(df, fdr_col = FDR, lfc_col = logFC, fdr = 0.05, lfc = 0)
}
\arguments{
\item{df}{dataframe of results. Must have columns containing significance values and log-fold changes.}

\item{fdr_col}{dataframe column (unquoted). Column of dataframe containing the significance level values.}

\item{lfc_col}{dataframe column (unquoted). Column of dataframe containing the lof-fold change values.}

\item{fdr}{numeric. FDR or significance value below which genes are considered significant.}

\item{lfc}{numeric. abs(log-fold change) value above which genes are considered significant.}
}
\value{
tibble of summarized results
}
\description{
Summarize a results data.frame. Return data.frame of counts of up/down/non-DE 
genes based on log-fold-change and significance values.
}
\examples{
library(coriell)


# Create a simulated results data.frame
# 50 will have logFC > 0
# 50 will have logFC < 0
# 25 of the up-regulated will have significant FDR values (< 0.0001)
# 25 of the down-regulated will have significant FDR values (< 0.0001)
df <- data.frame(feature_id = paste("gene", 1:100,sep = "."),
                 logFC = c(runif(50, min = -8, max = -0.01), 
                           runif(50, min = 0.01, max = 8)),
                 FDR = c(runif(25, min = 1e-12, max = 1e-4),
                         runif(25, min = 0.06, max = 1),
                         runif(25, min = 1e-12, max = 1e-4),
                         runif(25, min = 0.06, max = 1))
)

# view the results data.frame
head(df)

# summarize results
summarize_dge(df)

}
