% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_coriell}
\alias{theme_coriell}
\title{ggplot2 theme for coriell package}
\usage{
theme_coriell()
}
\description{
Based on \code{ggplot2::theme_light()} this theme will adjust the plot title,
legend position (bottom), axis titles, axis ticks, and strip text.
}
\examples{
library(ggplot2)

df <- data.frame(
  X = 1:100, 
  Y = rnorm(100), 
  C = sample(LETTERS[1:4], 100, replace = TRUE)
  )

# Default ggplot2
ggplot(df, aes(X, Y, color = C)) +
  geom_point() +
  labs(title = "ggplot2 default plot",
       subtitle = "subtitle for default plot")

# With theme_coriell()
ggplot(df, aes(X, Y, color = C)) +
  geom_point() +
  labs(title = "Plot with theme_coriell()",
       subtitle = "Subtitle for plot with theme_coriell()") +
   theme_coriell()

}
