% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tpm}
\alias{tpm}
\alias{tpm.default}
\alias{tpm.matrix}
\alias{tpm.data.frame}
\alias{tpm.SummarizedExperiment}
\alias{tpm.DelayedArray}
\title{Compute TPM normalized counts}
\usage{
tpm(x, l, ...)

\method{tpm}{default}(x, l, ...)

\method{tpm}{matrix}(x, l)

\method{tpm}{data.frame}(x, l)

\method{tpm}{SummarizedExperiment}(x, l, assay = "counts")

\method{tpm}{DelayedArray}(x, l)
}
\arguments{
\item{x}{matrix, numeric data.frame, or SummarizedExperiment object}

\item{l}{vector of feature lengths}

\item{assay}{if SummarizedExperiment, what assay to use. Default = "counts"}
}
\value{
matrix containing TPM normalized counts
}
\description{
Convert counts to TPM values. This function simply performs the computations
for creating TPM normalized counts given a count matrix and any vector of
lengths. This is useful for quickly computing scaling factors for arbitrary
features and lengths. For RNA-seq, it's better to use a transcript aligner
+ tximport to get TPMs. Refer to \url{https://bioconductor.org/packages/release/bioc/html/tximport.html}
for more information.
}
\examples{

# Generate some fake data with fake positive feature lengths
counts <- coriell::simulate_counts()$table
gene_lengths <- rbinom(nrow(counts), 1000, 0.9)

tpms <- tpm(counts, gene_lengths)

}
