% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore-df.R
\name{zscore_df}
\alias{zscore_df}
\title{Z-score a dataframe}
\usage{
zscore_df(df, by = "row")
}
\arguments{
\item{df}{data.frame or matrix of numeric columns}

\item{by}{a character string indicating which axis to scale by. Either "row" or "column". Default ("row")}
}
\value{
data.frame with either
}
\description{
Convenience function for scaling a dataframe by rows or columns.
}
\examples{
df <- data.frame(a = runif(10),
                 b = runif(10),
                 c = runif(10))
rownames(df) <- paste0("row", rownames(df))

zscore_df(df, by = "row")

}
