% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{exact_cor_test}
\alias{exact_cor_test}
\title{Perform and exact correlation test on every row of a matrix}
\usage{
exact_cor_test(X, y, ...)
}
\arguments{
\item{X}{numeric matrix or data.frame that can be converted to a numeric matrix}

\item{y}{numeric vector of data to correlate}

\item{...}{arguments passed to the `cor` function}
}
\value{
data.frame containing the original values in X along with columns
containing the correlation of Xi and Y, the empirical p-value of the
permutation test, and the FDR corrected empirical p-value
}
\description{
Perform an exact correlation permutation test. Computes correlation value for
every row of matrix X against all permutations of vector y.
}
\keyword{internal}
