% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{jackknifeSE}
\alias{jackknifeSE}
\title{Perform jackknife resampling on all columns of a SummarizeExperiment object}
\usage{
jackknifeSE(x, FUN, ...)
}
\arguments{
\item{x}{SummarizedExperiment object to perform jackknife resampling of columns on}

\item{FUN}{function to perform on each resample.}

\item{...}{Additional arguments passed to FUN}
}
\value{
List of results calculated by FUN on each resample of the SE object
}
\description{
This function provides a simple wrapper to perform jackknife resampling on
all columns of a SummarizedExperiment object and a return the calculated
results. This function is designed to be used to assess the robustness of 
p-value combination techniques of the included \code{meta_de()} function.
}
\examples{
# Define two differential expression dataset data.frames
exp1 <- data.frame(
  feature_id = c("geneA", "geneB", "geneC"),
  PValue = c(0.01, 0.5, 0.05),
  FDR = c(0.02, 0.5, 0.07),
  logFC = c(1.2, -2.5, 3.7),
  logCPM = c(12, 9, 0)
)

exp2 <- data.frame(
  feature_id = c("geneA", "geneB", "geneD"),
  PValue = c(0.07, 0.3, 0.8),
  FDR = c(0.08, 0.4, 1.0),
  logFC = c(1.5, -2.0, 3.0),
  logCPM = c(14, 10, 2)
)

# Combine into a single list
l <- list(experiment1 = exp1, experiment2 = exp2)

# Convert the data to a SummarizedExperiment
se <- dfs2se(l)

# Perform the jackknife using meta_de on each subset of the data
results <- jackknifeSE(se, \(x) meta_de(x, metapod::parallelWilkinson, min.prop = 0.5))

}
