% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_value}
\alias{map_value}
\title{Map a value in one range to a value in another}
\usage{
map_value(x, old_min, old_max, new_min, new_max)
}
\arguments{
\item{x}{numeric. Value to be transformed}

\item{old_min}{numeric. Minimum value of source range.}

\item{old_max}{numeric. Maximum value of source range.}

\item{new_min}{numeric. Minimum value of desired range.}

\item{new_max}{numeric. Maximum value of desired range.}
}
\value{
Value of x mapped to the desired range.
}
\description{
Map a value in one range to a value in another
}
\examples{
# Map values from the range 0-10 to the range 0-1
map_value(c(0, 1, 2, 3), 0, 10, 0, 1)
}
