% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{perm_cor}
\alias{perm_cor}
\title{Correlation Permutation Test Function}
\usage{
perm_cor(x, y, n_perm, cor_method)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{n_perm}{numeric. The number of permutations to perform}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of "pearson", "kendall", or "spearman": can be abbreviated.}
}
\value{
numeric. empirical p-value from permutation test.
}
\description{
Internal function for performing a correlation permutation test
}
\keyword{internal}
