% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup_profile.R
\name{pileup_profile}
\alias{pileup_profile}
\title{Return the pileup of reads over a set of GRanges}
\usage{
pileup_profile(
  bamfile,
  x,
  scan_bam_flag,
  sample_size = NULL,
  max_depth = 10000,
  min_base_quality = 13,
  min_mapq = 1,
  min_nucleotide_depth = 0
)
}
\arguments{
\item{bamfile}{Path to an indexed BAM file}

\item{x}{GRanges to calculate pileups over}

\item{scan_bam_flag}{An instance of Rsamtools::scanBamFlag() used to to create 
a parameter object influencing what fields and which records are imported 
from a (binary) BAM file}

\item{sample_size}{Number of records ranges to be randomly subsampled from x. 
Default NULL, no subsampling.}

\item{max_depth}{integer; maximum number of overlapping alignments considered 
for each position in the pileup. See Rsamtools::PileupParam()}

\item{min_base_quality}{integer; minimum ‘QUAL’ value for each nucleotide 
in an alignment. Use phred2ASCIIOffset to help translate numeric or 
character values to these offsets.}

\item{min_mapq}{integer; minimum ‘MAPQ’ value for an alignment to be included 
in pileup.}

\item{min_nucleotide_depth}{integer(1); minimum count of each nucleotide 
(independent of other nucleotides) at a given position required for said 
nucleotide to appear in the result.}
}
\value{
data.table
}
\description{
This function computes pileups for BAM files over a set of GRanges. It 
returns sum and average counts for all positions relative to the 5' end of 
the ranges. This function also allows for subsampling of ranges to speed up
calculations.
}
\details{
This function takes in a BAM file and computes a pileup over the given ranges. 
The data.table which gets returned contains the sum and average count of 
reads at each position relative to the 5' end of the input ranges. The user 
can also randomly subsample the input ranges to speed up computation. 
For both subsampling and pileup computation, ranges are split by positive and 
negative strand, subsampled at the same depth (if set), and positions in the 
final data.table are returned relative to the 5' start in the input ranges.
}
