% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Create boxplot from an expression matrix}
\usage{
plot_boxplot(x, metadata = NULL, fillBy = NULL, ...)
}
\arguments{
\item{x}{gene by sample matrix or data.frame of numeric values to be plotted.}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata
must match the colnames of the input matrix. Default NULL, each sample in the 
matrix will be plotted.}

\item{fillBy}{metadata column used to group and fill boxplots. Default NULL, 
each sample in the matrix will be plotted.}

\item{...}{Additional parameters passed to \code{ggplot2::geom_boxplot()}}
}
\value{
ggplot object
}
\description{
Create a boxplot of expression values for the given expression matrix.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the boxplot by sample
plot_boxplot(GSE161650_lc) + 
  theme_coriell()
  
# Plot the boxplot by Group in metadata
plot_boxplot(GSE161650_lc, metadata, fillBy = "Group") + 
  theme_coriell()
}
