% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density.R
\name{plot_density}
\alias{plot_density}
\title{Density plot of expression values}
\usage{
plot_density(x, metadata = NULL, colBy = NULL, ...)
}
\arguments{
\item{x}{gene by sample matrix or data.frame of numeric values to be plotted.}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata
must match the colnames of the input matrix. Default NULL, each sample in the 
matrix will be plotted.}

\item{colBy}{metadata column used to color density lines. Default NULL, 
each sample in the matrix will be plotted.}

\item{...}{Additional parameters passed to \code{ggplot2::geom_density()}}
}
\value{
ggplot object
}
\description{
Create a density of expression values for the given expression matrix.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the density by sample
plot_density(GSE161650_lc) + 
  theme_coriell()
  
# Color each sample by their Group in metadata
plot_density(GSE161650_lc, metadata, colBy = "Group") + 
  theme_coriell()
}
