% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_MD}
\alias{plot_MD}
\title{MD plot function}
\usage{
plot_MD(
  result,
  graph_title,
  fdr = 0.05,
  lfc = 0,
  up_alpha = 0.8,
  down_alpha = 0.8,
  non_alpha = 0.1
)
}
\arguments{
\item{result}{edgeR result object. Required.}

\item{graph_title}{Title of the plot. Required.}

\item{fdr}{Significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{Log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default(0)}

\item{up_alpha}{Alpha value of points for up-regulated genes. Default (0.8)}

\item{down_alpha}{Alpha value of points for down-regulated genes. Default (0.8)}

\item{non_alpha}{Alpha value of points for non-differentially expressed genes. Default (0.1)}
}
\value{
ggplot MD plot
}
\description{
Create an MD plot from an edgeR results object
}
