% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-md.R
\name{plot_md}
\alias{plot_md}
\title{Create an MD plot from expression data}
\usage{
plot_md(
  df,
  x = "logCPM",
  y = "logFC",
  sig_col = "FDR",
  lab = NULL,
  fdr = 0.1,
  lfc = 0,
  annotate_counts = TRUE,
  label_sig = FALSE,
  up_color = "red2",
  down_color = "royalblue2",
  nonde_color = "grey40",
  up_alpha = 1,
  down_alpha = 1,
  nonde_alpha = 1,
  up_size = 1,
  down_size = 1,
  nonde_size = 1,
  xmax_label_offset = 0.8,
  ymax_label_offset = 0.5,
  ymin_label_offset = 0.5,
  lab_size = 8,
  ...
)
}
\arguments{
\item{df}{dataframe containing log-fold-change, p-value, and logCPM columns.}

\item{x}{column in dataframe containing the logCPM data. Default ("logCPM")}

\item{y}{column in dataframe containing the log-fold-change values. Default ("logFC")}

\item{sig_col}{column in dataframe containing the results from significance testing. Default ("FDR")}

\item{lab}{column in dataframe containing the labels to plot if label_sig = TRUE. Default NULL}

\item{fdr}{numeric. Significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{numeric. Log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default(0)}

\item{annotate_counts}{TRUE/FALSE. Annotate the plot with the summarized gene counts}

\item{label_sig}{logical. Apply \code{ggrepel::geom_text_repel()} to significant DE genes. Default FALSE}

\item{up_color}{Point color of the up-regulated features. Default ("red2")}

\item{down_color}{Point color of the down-regulated features. Default ("royalblue2")}

\item{nonde_color}{Point color of the unperturbed features. Default ("grey40")}

\item{up_alpha}{Point alpha value of the up-regulated features. Default (1)}

\item{down_alpha}{Point alpha value of the down-regulated features. Default (1)}

\item{nonde_alpha}{Point alpha value of the unperturbed features. Default (1)}

\item{up_size}{Point size of the up-regulated features. Default (1)}

\item{down_size}{Point size of the down-regulated features. Default (1)}

\item{nonde_size}{Point size of the unperturbed features. Default (1)}

\item{xmax_label_offset}{numeric. Value between 0 and 1 inclusive. Controls the x-position of the count labels}

\item{ymax_label_offset}{numeric. Value between 0 and 1 inclusive. Controls the y-position of the 'up' count label}

\item{ymin_label_offset}{numeric. Value between 0 and 1 inclusive. Controls the y-position of the 'down' count label}

\item{lab_size}{numeric. If annotate_counts = TRUE specify the label size. Default = 8.}

\item{...}{Additional arguments passed to \code{ggrepel::geom_text_repel()}}
}
\value{
ggplot MD plot
}
\description{
Create an MD (MA) plot from a data.frame containing differential expression results.
}
\examples{
plot_md(GSE161650_de, fdr = 0.01, lfc = log2(2))
}
