% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-parallel.R
\name{plot_parallel}
\alias{plot_parallel}
\alias{plot_parallel.default}
\alias{plot_parallel.matrix}
\alias{plot_parallel.data.frame}
\alias{plot_parallel.SummarizedExperiment}
\title{Parallel coordinates plot of expression data}
\usage{
plot_parallel(x, ...)

\method{plot_parallel}{default}(x)

\method{plot_parallel}{matrix}(x, metadata = NULL, colBy = NULL, removeVar = NULL, ...)

\method{plot_parallel}{data.frame}(x, metadata = NULL, colBy = NULL, removeVar = NULL, ...)

\method{plot_parallel}{SummarizedExperiment}(x, assay = "counts", colBy = NULL, removeVar = NULL, ...)
}
\arguments{
\item{x}{gene by sample matrix or \code{SummarizedExperiment} object}

\item{...}{Additional parameters passed to \code{ggplot2::geom_line()}}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata}

\item{colBy}{metadata column used to color lines. Default NULL, every sample
will get its own color.}

\item{removeVar}{Remove this proportion of features based on the variance
across rows. Default NULL, all features are plotted.}

\item{assay}{assay of \code{SummarizedExperiment} object to be plotted. Default "counts".}
}
\value{
ggplot object
}
\description{
The parallel coordinates plot will display a line plot showing the expression
value for gene on the y-axis by each sample on the x-axis.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the PCP for each sample -- passing alpha value to geom_line()
plot_parallel(GSE161650_lc, alpha = 0.01) +
  theme_coriell()

# Plot the PCP by coloring each sample by Group from metadata
plot_parallel(GSE161650_lc, metadata, colBy = "Group", alpha = 0.01) +
  theme_coriell()
}
