% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-parallel.R
\name{plot_parallel}
\alias{plot_parallel}
\title{Parallel coordinates plot of expression data}
\usage{
plot_parallel(
  x,
  metadata = NULL,
  stat = "median",
  colBy = NULL,
  removeVar = NULL,
  ...
)
}
\arguments{
\item{x}{gene by sample matrix or data.frame of numeric values to be plotted.}

\item{metadata}{data.frame containing metadata per sample. rownames of metadata
must match the colnames of the input matrix. Default NULL, each sample in the 
matrix will be plotted.}

\item{stat}{Summary statistic to plot for each gene. One of c("mean", "median", "min", "max").
Default "median"}

\item{colBy}{metadata column used to color lines. Default NULL, every sample 
will get its own color.}

\item{removeVar}{If not NULL remove this proportion of features based on the 
variance across rows. Default NULL.}

\item{...}{Additional parameters passed to \code{ggplot2::geom_line()}}
}
\value{
ggplot object
}
\description{
The parallel coordinates plot will display a line plot showing the expression
value for gene by each sample or summarized expression value of a gene for the 
group given by the "colBy" argument on the y-axis and the sample or group 
label on the x-axis.
}
\examples{
# Create metadata for plotting
metadata <- data.frame(row.names = colnames(GSE161650_lc))
metadata$Group <- rep(c("DMSO", "THZ1"), each = 3)

# Plot the PCP for each sample -- passing alpha value to geom_line()
plot_parallel(GSE161650_lc, alpha = 0.01) + 
  theme_coriell()

# Plot the PCP by Group in metadata
plot_parallel(GSE161650_lc, metadata, colBy = "Group", alpha = 0.01) + 
  theme_coriell()
}
