% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-parallel2.R
\name{plot_parallel2}
\alias{plot_parallel2}
\title{Parallel coordinates plot}
\usage{
plot_parallel2(
  x,
  remove_var = 0.9,
  scale = TRUE,
  color = "black",
  alpha = 0.1,
  plot_title = NULL,
  x_label = NULL,
  y_label = NULL,
  ...
)
}
\arguments{
\item{x}{feature x sample matrix}

\item{remove_var}{numeric. What proportion of low variance features to remove from the matrix
before plotting. Default 0.9}

\item{scale}{bool. Should the data be standardized prior to plotting. Default TRUE}

\item{color}{character. Color of the lines. Default "black"}

\item{alpha}{numeric. Alpha value used to add transparency to lines. Default 0.1}

\item{plot_title}{character. title of the plot. Default NULL}

\item{x_label}{character. x-axis label}

\item{y_label}{character. y-axis label}
}
\value{
Parallel coordinate plot of scaled matrix data
}
\description{
Create a parallel coordinates (scaled expression data on y-axis, samples on x-axis)
for a matrix of data.
}
\examples{

plot_parallel2(GSE161650_lc, main = "10\% Most Variable Genes")

}
