% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-parallel2.R
\name{plot_parallel2}
\alias{plot_parallel2}
\title{Parallel coordinates plot}
\usage{
plot_parallel2(
  x,
  remove_var = 0.9,
  scale = TRUE,
  color = "black",
  alpha = 0.1,
  ...
)
}
\arguments{
\item{x}{feature x sample matrix}

\item{remove_var}{numeric. What proportion of low variance features to remove from the matrix
before plotting. Default 0.9}

\item{scale}{bool. Should the data be standardized prior to plotting. Default TRUE}

\item{color}{character. Color of the lines. Default "black"}

\item{alpha}{numeric. Alpha value used to add transparency to lines. Default 0.1}

\item{...}{Additional variables passed to \code{MASS:parcoord()}}
}
\value{
Parallel coordinate plot of scaled matrix data
}
\description{
Create a parallel coordinates (scaled expression data on y-axis, samples on x-axis)
for a matrix of data. This function calls \code{MASS::parcoord()} after optionally
removing the low variance features and scaling the data.
}
\examples{

plot_parallel2(GSE161650_lc, main = "10\% Most Variable Genes")

}
