% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create a volcano plot from expression data}
\usage{
plot_volcano(
  df,
  x = "logFC",
  y = "FDR",
  lab = NULL,
  fdr = 0.05,
  lfc = 0,
  label_sig = FALSE,
  annotate_counts = TRUE,
  up_color = "red2",
  down_color = "royalblue2",
  nonde_color = "grey40",
  up_alpha = 1,
  down_alpha = 1,
  nonde_alpha = 1,
  up_size = 1,
  down_size = 1,
  nonde_size = 1,
  up_shape = 16,
  down_shape = 16,
  nonde_shape = ".",
  xmin_label_offset = 0.8,
  xmax_label_offset = 0.8,
  ymax_label_offset = 0.9,
  lab_size = 6,
  lab_digits = 1,
  x_axis_limits = NULL,
  y_axis_limits = NULL,
  ...
)
}
\arguments{
\item{df}{dataframe containing columns with gene names, p-values, and log-fold changes.}

\item{x}{logFC column. Default (logFC)}

\item{y}{FDR or p-value column. Default (FDR)}

\item{lab}{column containing gene id or labels. Default (NULL)}

\item{fdr}{significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default (0)}

\item{label_sig}{TRUE/FALSE. apply ggrepel::geom_text_labels to significant DE genes.}

\item{annotate_counts}{TRUE/FALSE. Annotate the plot with the summarized gene counts}

\item{up_color}{Point color of the up-regulated features}

\item{down_color}{Point color of the down-regulated features}

\item{nonde_color}{Point color of the unperturbed features}

\item{up_alpha}{Point alpha of the up-regulated features}

\item{down_alpha}{Point alpha of the down-regulated features}

\item{nonde_alpha}{Point alpha of the unperturbed features}

\item{up_size}{Point size of the up-regulated features}

\item{down_size}{Point size of the down-regulated features}

\item{nonde_size}{Point size of the unperturbed features}

\item{up_shape}{Point shape of the up-regulated features}

\item{down_shape}{Point shape of the down-regulated features}

\item{nonde_shape}{Point shape of the unperturbed features}

\item{xmin_label_offset}{numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'down' counts}

\item{xmax_label_offset}{numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'up' counts}

\item{ymax_label_offset}{numeric. Value between 0 and 1 inclusive to control the y-position of the count labels.}

\item{lab_size}{numeric. Size of the label if annotate_counts = TRUE. Default 6.}

\item{lab_digits}{numeric. The number of digits used when rounding percentage values when annotate_counts=TRUE. Default (1)}

\item{x_axis_limits}{numeric vector of axis limits supplied to ggplot2::coord_cartesian(). Default (NULL)}

\item{y_axis_limits}{numeric vector of axis limits supplied to ggplot2::coord_cartesian(). Default (NULL)}

\item{...}{Additional arguments passed to \code{ggrepel::geom_text_repel}}
}
\value{
ggplot volcano plot
}
\description{
Create a volcano plot from a data.frame containing differential expression results.
}
\examples{
plot_volcano(GSE161650_de, fdr = 0.01, lfc = log2(2))
}
