% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot Function}
\usage{
plot_volcano(
  df,
  x = logFC,
  y = FDR,
  lab = feature_id,
  fdr = 0.05,
  lfc = 0,
  label_sig = FALSE
)
}
\arguments{
\item{df}{dataframe containing columns with gene names, p-values, and log-fold changes.}

\item{x}{logFC column. Default (logFC)}

\item{y}{FDR or p-value column. Default (FDR)}

\item{lab}{column containing gene id or labels. Default (NULL)}

\item{fdr}{significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default (0)}

\item{label_sig}{TRUE/FALSE. apply ggrepel::geom_text_labels to significant DE genes.}
}
\value{
ggplot volcano plot
}
\description{
Create a volcano plot from the given dataframe
}
\examples{
\dontrun{
library(edgeR)
library(coriell)

# create some fake data
x <- data.frame(
  ctl1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  ctl2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  row.names = paste0("gene", 1:1000)
)

# run edger pipeline
group <- factor(c(1, 1, 2, 2))
y <- DGEList(counts = x, group = group)
y <- calcNormFactors(y)
design <- model.matrix(~group)
y <- estimateDisp(y, design)

# To perform quasi-likelihood F-tests:
fit <- glmQLFit(y, design)
qlf <- glmQLFTest(fit, coef = 2)

# convert the results object to a dataframe
res_df <- edger_to_df(qlf)

# Create volcano plot
plot_volcano(res_df)
}

}
