% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot Function}
\usage{
plot_volcano(
  result,
  graph_title,
  fdr = 0.05,
  lfc = 0,
  label_sig = FALSE,
  pt_alpha = 0.25,
  pt_size = 2.25
)
}
\arguments{
\item{result}{edgeR results object. Required}

\item{graph_title}{Title of the plot. Required.}

\item{fdr}{Significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{Log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default (0)}

\item{label_sig}{bool. apply ggrepel::geom_text_labels to significant DE genes.}

\item{pt_alpha}{Alpha level of the points. Default (0.25)}

\item{pt_size}{Size of the points. Default (2.25)}
}
\value{
ggplot volcano plot
}
\description{
Create a volcano plot from an edgeR results object
}
\keyword{plot}
\keyword{plot,}
\keyword{volcano}
\keyword{volcano,}
