% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create a volcano plot from expression data}
\usage{
plot_volcano(
  df,
  x = "logFC",
  y = "FDR",
  lab = "feature_id",
  fdr = 0.1,
  lfc = 0,
  label_sig = FALSE,
  annotate_counts = TRUE,
  up_color = "red2",
  down_color = "red2",
  nonde_color = "grey40",
  up_alpha = 1,
  down_alpha = 1,
  nonde_alpha = 1,
  up_size = 1,
  down_size = 1,
  nonde_size = 1,
  xmin_label_offset = 0.5,
  xmax_label_offset = 0.5,
  ymax_label_offset = 0.8,
  lab_size = 10,
  ...
)
}
\arguments{
\item{df}{dataframe containing columns with gene names, p-values, and log-fold changes.}

\item{x}{logFC column. Default (logFC)}

\item{y}{FDR or p-value column. Default (FDR)}

\item{lab}{column containing gene id or labels. Default (NULL)}

\item{fdr}{significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)}

\item{lfc}{log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default (0)}

\item{label_sig}{TRUE/FALSE. apply ggrepel::geom_text_labels to significant DE genes.}

\item{annotate_counts}{TRUE/FALSE. Annotate the plot with the summarized gene counts}

\item{up_color}{Point color of the up-regulated features}

\item{down_color}{Point color of the down-regulated features}

\item{nonde_color}{Point color of the unperturbed features}

\item{up_alpha}{Point alpha of the up-regulated features}

\item{down_alpha}{Point alpha of the down-regulated features}

\item{nonde_alpha}{Point alpha of the unperturbed features}

\item{up_size}{Point size of the up-regulated features}

\item{down_size}{Point size of the down-regulated features}

\item{nonde_size}{Point size of the unperturbed features}

\item{xmin_label_offset}{numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'down' counts}

\item{xmax_label_offset}{numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'up' counts}

\item{ymax_label_offset}{numeric. Value between 0 and 1 inclusive to control the y-position of the count labels.}

\item{lab_size}{numeric. Size of the label if annotate_counts = TRUE. Default 10.}

\item{...}{Additional arguments passed to \code{ggrepel::geom_text_repel}}
}
\value{
ggplot volcano plot
}
\description{
Create a volcano plot from a data.frame containing differential expression results.
}
\examples{
\dontrun{
library(edgeR)
library(coriell)

# create some fake data
x <- data.frame(
  ctl1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  ctl2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  row.names = paste0("gene", 1:1000)
)

# run edger pipeline
group <- factor(c(1, 1, 2, 2))
y <- DGEList(counts = x, group = group)
y <- calcNormFactors(y)
design <- model.matrix(~group)
y <- estimateDisp(y, design)

# To perform quasi-likelihood F-tests:
fit <- glmQLFit(y, design)
qlf <- glmQLFTest(fit, coef = 2)

# convert the results object to a dataframe
res_df <- edger_to_df(qlf)

# Create volcano plot
plot_volcano(res_df)
}

}
