% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{quickmap}
\alias{quickmap}
\title{Quick Heatmap with Sensible Defaults for RNA-seq}
\usage{
quickmap(mat, ...)
}
\arguments{
\item{mat}{numeric matrix to be passed onto pheatmap function}

\item{...}{args to be passed to pheatmap function}
}
\value{
pheatmap object. See ?pheatmap::pheatmap for details
}
\description{
Generate a heatmap using pheatmap with sensible defaults for RNA-seq.
Utilizes a default blue to red diverging color palette, removes rownames
and borders, clusters rows and columns, sets row-wise clustering is euclidean
column-wise clustering is correlation with clustering method set to complete.
Display font size is reduced and column labels are set to an angle. Defaults can
be overridden and/or any additional arguments can be passed to the pheatmap
function via ...
}
