% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{quickmap}
\alias{quickmap}
\title{Heatmap with sensible defaults for RNA-seq expression data}
\usage{
quickmap(x, metadata = NULL, remove_var = NULL, ...)
}
\arguments{
\item{x}{numeric matrix to be passed onto pheatmap function}

\item{metadata}{column metadata to add to heatmap annotation. Default NULL}

\item{remove_var}{Remove this proportion of features based on variance 
across rows. Default NULL.}

\item{...}{args to be passed to \code{pheatmap()} function}
}
\value{
pheatmap object. See \code{?pheatmap::pheatmap()} for details
}
\description{
Generate a heatmap using \code{pheatmap} with sensible defaults for RNA-seq.
}
\details{
The default arguments to \code{pheatmap::pheatmap()} are:
\itemize{
  * \code{scale = "row"}
  * \code{show_rownames = FALSE}
  * \code{border_col = NA}
  * \code{cluster_rows = TRUE}
  * \code{cluster_cols = TRUE}
  * \code{color = colorRampPalette(c("dodgerblue3", "grey99", "firebrick3"))(30)}
  * \code{treeheight_row = 0}
  * \code{clustering_distance_rows = "correlation"}
  * \code{clustering_distance_cols = "euclidean"}
  * \code{clustering_method = "ward.D2"}
  * \code{angle_col = 315}
}
}
\examples{

# Display heatmap of scaled count data and add title to the plot
quickmap(GSE161650_lc, main = "THZ1 vs DMSO")

# Override quickmap defaults by supplying additional args
quickmap(GSE161650_lc, cluster_rows = FALSE, cluster_cols = FALSE)

}
