% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{quickmap}
\alias{quickmap}
\title{Heatmap with sensible defaults for RNA-seq expression data}
\usage{
quickmap(mat, diverging_palette = TRUE, ...)
}
\arguments{
\item{mat}{numeric matrix to be passed onto pheatmap function}

\item{diverging_palette}{logical. Default(TRUE). Sets the color scale to a diverging palette (blue -> white -> red). If FALSE, set the
color to a continuous color palette \code{viridis::magma}, useful for un-scaled expression data.}

\item{...}{args to be passed to \code{pheatmap} function}
}
\value{
pheatmap object. See \code{?pheatmap::pheatmap} for details
}
\description{
Generate a heatmap using \code{pheatmap} with sensible defaults for RNA-seq.
}
\details{
The default arguments to \code{pheatmap::pheatmap()} are:
\itemize{
\itemize{
\item \code{scale = "row"}
\item \code{show_rownames = FALSE}
\item \code{border_col = NA}
\item \code{cluster_rows = TRUE}
\item \code{cluster_cols = TRUE}
\item \code{color = colorRampPalette(c("dodgerblue3", "grey99", "firebrick3"))(50)} for diverging_palette = TRUE
\item \code{color = rev(viridisLite::magma(n = 50))} for diverging_palette = FALSE
\item \code{treeheight_row = 0}
\item \code{clustering_distance_rows = "correlation"}
\item \code{clustering_distance_cols = "euclidean"}
\item \code{clustering_method = "ward.D2"}
\item \code{angle_col = 315}
}
You can pass in additional arguments or simply override the defaults as well.
}
}
\examples{
# generate fake count data
X <- coriell::simulate_counts(n_genes = 100)$table

# display heatmap of scaled count data and add title to the plot
quickmap(X, main = "Control vs Treatment")
}
