% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{sample_n_random_cor}
\alias{sample_n_random_cor}
\title{Generate a null distribution of correlation values}
\usage{
sample_n_random_cor(X, y, n = 10000, ...)
}
\arguments{
\item{X}{numeric matrix or data.frame that can be converted to numeric matrix.}

\item{y}{numeric vector. Numeric vector of values used to correlate with each row of df}

\item{n}{integer. Number of random correlations to return. Default (1e4)}

\item{...}{Additional arguments passed to `cor` function}
}
\value{
numeric vector of correlation values. Vector can contain NAs if row variance == 0.
}
\description{
Selects n random rows from a numeric matrix with replacement. For each random
row, permute vector y and perform correlation.
}
\examples{
# generate example data
X <- matrix(runif(1e3 * 6), nrow = 1e3, ncol = 6)
y <- 1:6
dimnames(X) <- list(paste("feature", 1:1e3, sep = "."), paste("sample", 1:6, sep = "."))

# sample random correlations from permuted data
null_dist <- sample_n_random_cor(X, y, n = 1e3, method = "spearman")

head(null_dist)
}
