% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-counts.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Simulate RNASeq Count Data}
\usage{
simulate_counts(
  n_genes = 1000,
  n_de = 100,
  n_samples = 6,
  groups = c("ctl", "trt"),
  de_group = "trt",
  mu = 10,
  phi = 0.1,
  count_offset = 25
)
}
\arguments{
\item{n_genes}{integer. The number of gene to simulate}

\item{n_de}{integer. The number of genes to be differentially expressed}

\item{n_samples}{integer. The number of samples to simulate}

\item{groups}{character vector. Character vector specifying the names of the groups to simulate}

\item{de_group}{character. The name of the group to be differentially expressed}

\item{mu}{numeric. Mean value of counts}

\item{phi}{numeric. Dispersion parameter}

\item{count_offset}{numeric. Count to be added to the de_group}
}
\value{
list containing the count matrix and a vector indicating which rows were modified.
}
\description{
Simulate an RNASeq count matrix. Adapted from edgeR documentation.
}
\examples{
# simulate counts using default parameters
sim <- simulate_counts()

# extract the count matrix from the simulation result
counts <- sim$table

# show rows where counts were modified
counts[sim$de_genes, ]
}
