% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-dge.R
\name{summarize_dge}
\alias{summarize_dge}
\title{Summarize RNA-seq expression results}
\usage{
summarize_dge(df, fdr_col = "FDR", lfc_col = "logFC", fdr = 0.05, lfc = 0)
}
\arguments{
\item{df}{dataframe of results. Must have columns containing significance values and log-fold changes.}

\item{fdr_col}{Column name of the data.frame containing the significance level values.}

\item{lfc_col}{Column name of data.frame containing the log-fold change values.}

\item{fdr}{numeric. FDR or significance value below which genes are considered significant.}

\item{lfc}{numeric. abs(log-fold change) value above which genes are considered significant.}
}
\value{
data.frame with columns describing the count and percentages of up/down/unperturbed genes
}
\description{
Summarize a results data.frame. Return data.frame of counts of up/down/non-DE 
genes based on log-fold-change and significance values.
}
\examples{
summarize_dge(GSE161650_de)
}
